requirements.txt
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
enum-compat==0.0.2
filetype==1.0.0
gcloud==0.18.3
googleapis-common-protos==1.5.3
greenlet==0.4.12
httplib2==0.10.3
idna==2.6
image==1.5.17
oauth2client==4.1.2
olefile==0.44
Pillow==4.3.0
protobuf==3.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyJWT==1.5.3
pytz==2017.3
requests==2.18.4
rsa==3.4.2
sheetsu==1.0.1
six==1.11.0
urllib3==1.22
